# Conflicts:
#	com.mire.cinema/src/main/java/com/mire/cinema/controller/ItemGiftCardController.java

+++ com.mire.cinema/src/main/java/com/mire/cinema/service/MemberServiceImpl.java
@@ -1,6 +1,5 @@
 package com.mire.cinema.service;
 
-import org.apache.ibatis.javassist.runtime.DotClass;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.config.annotation.authentication.builders.AuthenticationManagerBuilder;
 import org.springframework.security.core.Authentication;
@@ -10,12 +9,12 @@ import org.springframework.transaction.annotation.Transactional;
 import com.mire.cinema.domain.member.Member;
 import com.mire.cinema.domain.member.MemberDTO;
 import com.mire.cinema.domain.member.TokenDTO;
+import com.mire.cinema.exception.ErrorMsg;
 import com.mire.cinema.repository.MemberMapper;
 import com.mire.cinema.security.JwtTokenProvider;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-
 
 @Slf4j
 @Service
@@ -35,21 +34,20 @@ public class MemberServiceImpl implements MemberService {
 
 	@Override
 	public Member findMember(String memberId) {
-			
+
 		Member member = memberMapper.selectMember(memberId);
-		
+
 		return member;
 	}
 
 	@Override
 	public void modifyMember(MemberDTO.Update dto) {
 		Member member = memberMapper.selectMember(dto.getMemberId());
+
 		if (member == null || !member.getMemberPasswd().equals(dto.getMemberPasswd())) {
-			throw new IllegalArgumentException("회원님의 비밀번호 정보가 일치하지 않습니다.");
+			throw new IllegalArgumentException(ErrorMsg.USERINFO);
 
 		}
-
-		
 
 		memberMapper.updateMember(dto);
 
@@ -58,23 +56,35 @@ public class MemberServiceImpl implements MemberService {
 	@Override
 	public void removeMember(String memberId) {
 
+		Member member = findMember(memberId);
+
+		if (member == null) {
+			System.out.println("예외호출");
+			throw new IllegalArgumentException(ErrorMsg.USERINFO);
+		}
+
 		memberMapper.deleteMember(memberId);
 
 	}
 
 	@Override
 	public TokenDTO loginMember(MemberDTO.Login dto) {
-		System.out.println(dto);
-		UsernamePasswordAuthenticationToken authenticationToken = new UsernamePasswordAuthenticationToken(dto.getMemberId(), dto.getMemberPasswd());
-		
+		Member member = memberMapper.selectMember(dto.getMemberId());
+
+		if (member == null || !member.getMemberPasswd().equals(dto.getMemberPasswd())) {
+			throw new IllegalArgumentException(ErrorMsg.USERINFO);
+
+		}
+
+		UsernamePasswordAuthenticationToken authenticationToken = new UsernamePasswordAuthenticationToken(
+				dto.getMemberId(), dto.getMemberPasswd());
+
 		Authentication authentication = authenticationManagerBuilder.getObject().authenticate(authenticationToken);
-		
+
 		TokenDTO tokenDTO = jwtTokenProvider.createToken(authentication);
-	
-		
-		
+
 		return tokenDTO;
-		
+
 	}
 
 }
